feat: prototype of iris upgrade protocol
The iris upgrade protocol is in the crate `iris-mpc-upgrade` and lives
in the `upgrade` module.
A version showing its use using a simple TCP connection setup between
servers and clients is included in `src/bin`.

As part of this commit, I moved the GPU stuff into a sub-crate in the
new workspace.