Change HNSW PointID to type alias of u32 instead of usize tuple struct
PointID type is load-bearing for HNSW index data representation as it is used
in representing edges of the HNSW graph, so inefficiencies in the
representation have a large effect on the overall memory footprint of the
index.  32-bit indices support up to around 4.3B entries, which is large enough
for the currently planned scope of the iris uniqueness database.