Implement batch insertions for HNSW sorted lists (#1053)
Implementation of non-adaptive sorting networks and Batcher odd-even
merge sort.  Includes a generalized construction for the Batcher network
which supports sorting for input lists with a pre-sorted initial segment and
a total length which is not equal to a power of 2.

This batched insertion is integrated with the `SortedNeighborhood` struct
used in HNSW graphs via a new `insert_batch` method.  To support batched
distance comparisons, the interface for the `less_than_batch` function call
of the `VectorStore` trait is modified to accept a general list of `DistanceRef`
tuples for the comparison.  The HNSW search algorithm does not yet make
use of this batched insertion mode.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>