Implement batched HNSW distance comparisons (#1086)
* Implement greedy and batched search modes for HNSW layer search

* Change SortingNetwork struct to SwapNetwork to reflect usage beyond sorting

* Implement swap network for parallelized search for min list element

* Bump version for `ring` subdependency

Update needed in response to advisory RUSTSEC-2025-0009

* Ignore two security advisories for unmaintained dependencies

RUSTSEC-2024-0436 is for the `paste` crate, which is a deep dependency
of several AWS and SQLX related crates which is not easily extracted from
the codebase.

RUSTSEC-2025-0012 is for the `backoff` crate for exponential backoff used
in our gRPC implementation of `iris-mpc-cpu`.  This advisory is being
ignored only temporarily/briefly to support continued development, but
PR 1119 will eliminate the `backoff` dependency when merged.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>