Implement parallel MPC execution for HNSW graph insertion (#1159)
Adds support for parallel execution of binary insertion of a new node into the
neighbors' neighbor lists.

A new `BinarySearch` struct is implemented which keeps track of ongoing state
of a binary search procedure and provides the next array index which needs to
be compared against the query. This struct is used in `HnswSearcher::insert_prepare`
to execute binary insertion in parallel using one call to `less_than_batch` per
binary search step, or about 10 calls in total at realistic parameter sizes.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>