Implement seeded hnsw randomness for tests (#1249)
Adds a configuration parameter for an explicit base seed to be used for
HNSW graph generation. This seed is used to derive individual RNGs
associated with HawkSession structs, constructed by hashing this base
seed, the StoreId, and the SessionId into a final u64 seed. (Note this
construction needs to produce statistical randomness, but does not
need to be cryptographically secure, so a u64 seed suffices.) If no
explicit base seed is specified in the server Config (the hnsw_prng_seed
parameter is set to None) then the default behavior is used which
instead chooses a random seed per session agreed upon by the server
parties.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>
Co-authored-by: Mark A. Greenslade <mark@inversed>