Refactor HNSW server_main into smaller function calls (#1266)
This PR restructures the primary HNSW server_main function by
pulling out logical components of the server functionality into
individual functional calls. This restructuring is in preparation for
implementing the "genesis" indexer, which will be used to initially
construct an HNSW graph from pre-existing secret shared iris
shares, and which will share much (but not all) of the functionality
of server_main.

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>
Co-authored-by: Mark A. Greenslade <mark@inversed.tech>