Implement database persistence for genesis indexer (#1367)
* Implement persistence of batch processing results to database

* Change terminology "height" to "id" to reduce ambiguity

Existing terminology for how much of the iris db has been indexed used
the term "height" to represent the iris serial id of indexation.  This
is a bit ambiguous because the indexer skips some irises, so by the name
alone it is not clear if height refers to "the number of irises which
have been indexed" or "the last serial id which has been indexed".  The
latter meaning is intended, so using variable names with "id" rather
than "height" eliminates this ambiguity.

* Rework sanity checks for batches in genesis indexer main loop