Revise Postgres schemas to represent ids consistently (#1417)
This PR modifies the Postgres schemas for the `hawk_graph_links` and
`hawk_graph_entry` tables to represent ids consistently with the representation
used in the irises table, as `serial_id: bigint` and `version_id: smallint`
values. It additionally modifies the GraphOps implementations to specialize to
VectorStore structs with `VectorRef = VectorId`, and to interact properly with
the new schema.

The updated schema also adds explicit range constraints on column values in the
graph tables, and changes the `layer` and `graph_id` column types from integer
to smallint.