[POP-2533] Implement processing of modifications in delta (#1432)
PR adds support for processing modifications table entries persisted by the
production nodes since the last run of the genesis indexer. The approach is to
turn the genesis JobRequest and JobResult structs into enums with separate
variants for indexing of batches and for processing modification table entries.

The current version emulates the main server behavior for reauth and reset type
transactions, which is to simply re-insert the current newest entry into the
graph, without doing any pruning of old versions still existing in the graph. We
will need to modify this behavior later to do deletion of the existing graph
nodes and, very rarely, reassignment of the graph entry point.