Add genesis test cases 102 and 103 (#1563)
This PR implements the genesis 102 and 103 tests (POP-2713 and POP-2715),
provides support for initializing postgres modifications for e2e tests
(POP-2704), and begins implementing the genesis 104 test, which still needs
support for modifications to be added to the plaintext genesis simulator in
order to execute assertions over the final state.

The PR additionally adds a README.md to the genesis test runners submodule to
describe the functions of the different tests, and provides a utility to more
easily initialize plaintext genesis states for simulation.