[POP-2714] Implement delta protocol for plaintext genesis indexer (#1569)
This PR adds the genesis delta sub-protocol to the plaintext genesis
implementation, providing support for processing modifications present in the
source database. It additionally fixes a small bug in plaintext state
initialization, where the in-memory iris stores were initialized from the
destination database, but should instead be initialized from the source
database.