[POP-2718] Implement genesis 105 e2e test (#1579)
This PR extends the functionality for initializing and updating modifications
data in a test Postgres configuration and in the plaintext genesis state struct,
and implements the genesis 105 e2e test exercising mixed handling of
modifications initialized before a run, and then updated prior to a second run
of the genesis protocol. The PR also fixes some small incorrect edge case
behavior in the plaintext implementation of the genesis indexer (delta
subprotocol), so that the last_indexed_modification_id is properly updated in
the presence of non-finalized modifications.