Implement genesis 106 e2e test (#1580)
This PR implements the final main genesis e2e test, 106, which combines both
deletions and inline modifications during execution. This PR also tidies up the
README entries for tests 105 and 106, and adds functionality for comparing the
vector ids of irises present in the node CPU databases and the plaintext genesis
state struct.