[POP-3034] Implement HNSW layer search with batched node openings (#1738)
The existing implementation of batched HNSW layer search provides batching of
all basic distance comparison operations (e.g. filtering of visited nodes
against the current worst candidate node, and insertion of filtered nodes into
the candidate neighborhood), but processes node openings one at a time. This
means that the distances of neighbors of individual candidate nodes are computed
in a batch, but that there is no mechanism for batching together the openings of
multiple candidate nodes at once. This is more of an issue now because the
distance evaluation phase makes up a majority of the overall execution round
complexity.

This PR implements a new version of batched HNSW layer search which handles
batching at the node opening level: as before, an ongoing estimate of the "rate
of insertions of visited nodes into the candidate neighborhood" is maintained,
and multiple unopened nodes are opened as a batch, the number of which is chosen
so that the batch is expected to insert a fixed constant number of nodes per
iteration. This simplifies the overall procedure quite significantly, as
individual queues are not required for the separate filtering and opening
phases, since the selection of a number of nodes to open already provides rate
calibration. This additionally eliminates the "cleanup" phase of operation in
the previous version.

See `iris_mpc_cpu::hnsw::searcher::layer_search_batched_v2` for additional
details.