Fix bugs in apply_oblivious_swap_network
Two bugs were present in the existing implementation.  First, the
comparisons were reversed to allow a less-than-or-equal comparison
from a strict less-than primitive, but the bits were not subsequently
complemented to change a d2 < d1 to d1 <= d2.  Second, the distances
used for comparisons at each layer of a swap network were the original
distances input into the function, but should be the updated distances
permuted by the prior layers of the swap network.