[POP-3273] Implement fully batched neighborhood compaction (#1853)
* Implement fully batched neighborhood compaction

* Fix bugs in apply_oblivious_swap_network

Two bugs were present in the existing implementation.  First, the
comparisons were reversed to allow a less-than-or-equal comparison
from a strict less-than primitive, but the bits were not subsequently
complemented to change a d2 < d1 to d1 <= d2.  Second, the distances
used for comparisons at each layer of a swap network were the original
distances input into the function, but should be the updated distances
permuted by the prior layers of the swap network.

* Bump ampc-common dependencies to current main branch commit

* Fix import changed from ampc-common version update

* Improve efficiency of swap network processing

* Fixes in response to PR comments

* Port cargo-deny exception for RUSTSEC-2025-0134 from main

* Move main compact_neighborhood_batch logic to Aby3Store impl

* Add metric and logging for compacting neighborhoods count